Added: - Ability to delete files, - blocking the deletion of system-critical files, - ability to create new files (dir /opt/etc/nfqws), - validation of names for newly created files, - check for name conflicts when creating new files, - custom syntax highlighting for nfqws configs, - syntax highlighting toggle from shell to custom for nfqws, - updated color palette for dark and light themes

+++ index.php
@@ -81,7 +81,13 @@ function getLogContent(string $filename, $path = ROOT_DIR . '/var/log'): string 
 function saveFile(string $filename, string $content, $path = ROOT_DIR . '/etc/nfqws') {
     $filename = basename($filename);
     $file = $path . '/' . $filename;
-    return file_exists($file) && file_put_contents($file, normalizeString($content)) !== false;
+
+    $protected = ['nfqws.conf', 'user.list', 'exclude.list', 'auto.list', 'ipset.list', 'ipset_exclude.list', 'nfqws.log'];
+    if (!file_exists($file) && in_array($filename, $protected, true)) {
+        return false;
+    }
+
+    return file_put_contents($file, normalizeString($content)) !== false;
 }
 
 function saveLog(string $filename, string $content, $path = ROOT_DIR . '/var/log') {
@@ -91,6 +97,10 @@ function saveLog(string $filename, string $content, $path = ROOT_DIR . '/var/log
 function removeFile(string $filename, $path = ROOT_DIR . '/etc/nfqws') {
     $filename = basename($filename);
     $file = $path . '/' . $filename;
+    $protected = ['nfqws.conf', 'user.list', 'exclude.list', 'auto.list', 'ipset.list', 'ipset_exclude.list', 'nfqws.log'];
+    if (in_array($filename, $protected, true)) {
+        return false;
+    }
     if (file_exists($file)) {
         return unlink($file);
     } else {
